// packages/youtube-mcp/src/api/types.ts
export type YouTubeApiPlaylistItem = {
  kind?: string;
  etag?: string;
  id: string;
  snippet?: {
    title?: string;
    description?: string;
    channelId?: string;
    channelTitle?: string;
    publishedAt?: string;
    thumbnails?: Record<
      string,
      { url: string; width?: number; height?: number }
    >;
    tags?: string[];
    defaultLanguage?: string;
  };
  status?: {
    privacyStatus?: string;
  };
  contentDetails?: {
    itemCount?: number;
  };
};

export type YouTubeApiPlaylistItemResource = {
  kind?: string;
  etag?: string;
  id: string;
  snippet?: {
    title?: string;
    description?: string;
    channelId?: string;
    channelTitle?: string;
    publishedAt?: string;
    thumbnails?: Record<
      string,
      { url: string; width?: number; height?: number }
    >;
    playlistId?: string;
    position?: number;
    resourceId?: {
      kind?: string;
      videoId?: string;
    };
    videoOwnerChannelTitle?: string;
    videoOwnerChannelId?: string;
  };
  contentDetails?: {
    videoId?: string;
    startAt?: string;
    endAt?: string;
    note?: string;
    videoPublishedAt?: string;
  };
};
// ...
export type YouTubeApiResponse<T> = {
  kind?: string;
  etag?: string;
  nextPageToken?: string;
  prevPageToken?: string;
  pageInfo?: {
    totalResults?: number;
    resultsPerPage?: number;
  };
  items?: T[];
};
// ...
export type PlaylistDetails = {
  id: string;
  title: string;
  description: string;
  channelId: string;
  channelTitle: string;
  publishedAt: string;
  thumbnails: Record<string, { url: string; width?: number; height?: number }>;
  itemCount: number;
  privacy: string;
};

export type PlaylistItem = {
  id: string;
  title: string;
  description: string;
  videoId: string;
  channelId: string;
  channelTitle: string;
  playlistId: string;
  position: number;
  publishedAt: string;
  thumbnails: Record<string, { url: string; width?: number; height?: number }>;
};